Fix for ASPL-944
As discussed, I have removed Risk part and the divider line from list_apps response to give us more characters. After some testing the limit I have put to 30 apps to be one the safer side as Slack has constraints/limit on the number of characters that can be posted in a message.